+++ app/Http/Requests/User/StoreUserRequest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Http\Requests\Admin\User;
+namespace App\Http\Requests\User;
 
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateUserRequest extends FormRequest
+class StoreUserRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -23,12 +23,12 @@ class UpdateUserRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'name' => "required|string|max:255|unique:users,name,{$this->user->id}",
+            'name' => 'required|string|max:255|unique:users,name',
             'password' => 'required|string|max:255',
-            'email' => "required|email:rfc,dns|max:255|unique:users,email,{$this->user->id}",
+            'email' => 'required|email:rfc,dns|max:255|unique:users,email',
             'role' => 'required|integer',
             'status' => 'required|integer',
-            'full_name' => 'string|max:255|nullable'
+            'full_name' => 'string|max:255|nullable',
         ];
     }
